+++ src/App.tsx
@@ -3,7 +3,7 @@ import MobileLayout from './components/common/MobileLayout';
 import reset from './global/style/reset';
 import { theme } from './global/style/theme';
 import ShopPage from './pages/shop/ShopPage';
-import CartProvider from './context/CartContext';
+import CartProvider from './context/CartProvider';
 
 function App() {
   return (
